+++ src/cores/dataunit.py
@@ -1,19 +1,19 @@
 
 from typing import Any, Self
+from dataclasses import replace, asdict
+from datetime import datetime
 from pandas.core.api import DataFrame as DataFrame
 from interfaces.dataunit_interface import DataUnitInterface
+from cores.metas.dataunit_meta import DataUnitMeta
 
 class DataUnit(DataUnitInterface):
     
-    def __init__(self, name: str, df: DataFrame, metadata: dict[str, Any]) -> None:
+    def __init__(self, df: DataFrame, metadata: DataUnitMeta) -> None:
         """
         DataUnitの初期化。空のnameやDataFrameは禁止
         """
-        if not name:
-            raise ValueError("name must not be empty")
         if df.empty:
             raise ValueError("df must not be empty")
-        self._name = name
         self._df = df
         self._metadata = metadata
 
@@ -22,7 +22,7 @@ class DataUnit(DataUnitInterface):
         """
         getter ユニットの名前を返す
         """
-        return self._name
+        return self._metadata.name
 
     @property
     def df(self) -> DataFrame:
@@ -32,7 +32,7 @@ class DataUnit(DataUnitInterface):
         return self._df
 
     @property
-    def metadata(self) -> dict[str, Any]:
+    def metadata(self) -> DataUnitMeta:
         """
         getter ユニットのmetadataを返す
         """
@@ -42,21 +42,29 @@ class DataUnit(DataUnitInterface):
         """
         immutable setter 新しいユニット名をもった新規ユニットを返す
         """
-        return self.__class__(new_name, self.df, self.metadata)
+        new_metadata = replace(self._metadata, name=new_name)
+        new_namedata = replace(new_metadata, created_at = datetime.now())
+        return self.__class__(self.df, new_metadata)
 
     def with_updated_df(self, new_df: DataFrame) -> Self:
         """
         immutable setter 新しいDataFrameをもった新規ユニットを返す 
         """
-        return self.__class__(self.name, new_df, self.metadata)
+        return self.__class__(new_df, self.metadata)
 
     def with_update_metadata(self, key: str, value: Any) -> Self:
         """
         immutable setter 新しいメタデータを持った（更新した）新規ユニットを返す
+        いらないかもpendings
         """
-        new_metadata = self.metadata.copy()
-        new_metadata[key] = value
-        return self.__class__(self.name, self.df, new_metadata)
+        if not hasattr(self._metadata, key):
+            raise AttributeError('metadata objects dont have "{key}" field')
+        # asdictを用いてmetadataを辞書型に変換
+        metadata_dict = asdict(self._metadata)
+        
+        metadata_dict[key] = value
+        new_metadata = self._metadata.__class__(**metadata_dict)
+        return self.__class__(self.df, new_metadata)